requirements/dev.txt
-r base.txt

# Core requirements

Werkzeug==0.11.9
flake8==2.3.0
mock==1.0.1
django-debug-toolbar==1.5.0

# Addons requirements
